pyproject.toml
[project]
name = "thought-foundry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.83.0",
    "pyyaml>=6.0.3",
]
